==> analyses/1_exploration/markdowns/1_googledraw_cues.Rmd <==
---
title: googledraw cues
subtitle: 
author: Molly Lewis 
date: "`r Sys.Date()`"
output: 
  html_document:
    number_sections: false
    theme: cerulean
    toc_float: true
    code_folding: hide
---
  
******
******

```{r setup, include = F}
rm(list=ls())

# load packages
library(tidyverse) 
library(knitr)
library(rmarkdown)
library(broom)
library(langcog)
library(magrittr)

opts_chunk$set(echo = T, message = F, warning = F, 
               error = F, tidy = F,  cache = F, fig.height = 4)
```  

```{r}

arrange(concreteness, Conc.M) %>% 
  slice(200:400) %>% 
  as.data.frame()
```

```{r}
cues = read.table("../../data/supplementary_data/cues.txt", sep = "\n", header = FALSE) %>%
  rename(cue = V1)

concreteness <- read_csv("../../data/supplementary_data/brysbaert_corpus.csv")

cues %<>% 
  left_join(concreteness %>% select(Word, Conc.M, SUBTLEX), 
            by = c("cue"= "Word"))

ggplot(cues, aes(x = Conc.M)) +
  geom_histogram() +
  xlim(1,5) +
  theme_bw()

ggplot(cues, aes(x = log(SUBTLEX))) +
  xlab("log word frequency") +
  geom_histogram() +
  theme_bw()

extreme.cues = cues %>% 
  mutate(conc.bin = cut_number(Conc.M, n = 6, labels= FALSE)) %>%
  filter(conc.bin == 6 | conc.bin == 1)

write.csv(extreme.cues, "extreme_cues.csv")

cues %>%
  arrange(conc.bin) %>%
  kable()
```
