==> analyses/1_exploration/markdowns/7_dyadic_variables.Rmd <==
---
title: Explore country dyadic variables
subtitle: 
author: Molly Lewis 
date: "`r Sys.Date()`"
output: 
  html_document:
    number_sections: false
    theme: cerulean
    toc_float: true
    code_folding: hide
---
  
******
******

```{r setup, include = F}
rm(list=ls())

# load packages
library(knitr)
library(rmarkdown)
library(langcog)
library(tidyverse)
library(stringr)
library(purrr)
library(corrr)
library(countrycode)
library(forcats)

opts_chunk$set(echo = T, message = F, warning = F, 
               error = F, tidy = F,  cache = F, fig.height = 7, fig.width = 9)
``` 

Read in files
```{r}
geo <- read_csv("../../data/supplementary_data/cultural_sim_measures/geo/all_google_geo_dists.csv") %>%
  select(-capital_dist_meters)
lang <- read_csv("../../data/supplementary_data/cultural_sim_measures/lang/all_google_lang_dists.csv")
events<- read_csv("../../data/supplementary_data/cultural_sim_measures/events/all_google_event_dists.csv")
```

These are our variables: 
"centroid_dist_meters"  
"wals_euclidean_dist"/"asjp_dist"    
"log_normalized_n_events_all" 
"lop_normalized_mean_imports_dollars"
   
```{r}
get_unique_relation_id <- function (x, y){
  pairs = c(x, y)
  ordered = order(pairs)
  paste0(pairs[ordered[1]], pairs[ordered[2]])
}

all_dyadic2 <- events %>%
  full_join(geo)  %>%
  full_join(lang) %>%
  rowwise() %>%
  mutate(all_codes = get_unique_relation_id(country_code_1, country_code_2)) %>%
  ungroup() %>%
  select(all_codes, everything())

# this just cause there's a weird bug
#write_csv(all_dyadic2, "../../data/supplementary_data/cultural_sim_measures/events/all_dyadic2_temp.csv")

all_dyadic2 <- read_csv("../../data/supplementary_data/cultural_sim_measures/events/all_dyadic2_temp.csv") %>%
  mutate(all_codes = as.factor(all_codes))

is.na(all_dyadic2) <- do.call(cbind, lapply(all_dyadic2, is.infinite))

dyadic_clean <- all_dyadic2 %>%
  group_by(all_codes) %>%
  slice(1)  %>%
  ungroup()
```

## Histograms

Summary of dyadic measures:
```{r, fig.height = 6}
summary(dyadic_clean)

dyadic_clean %>%
  gather("measure", "value" , 4:8)  %>%
  ggplot(aes(x = value, fill = measure)) +
  geom_histogram() +
  facet_wrap(~measure, scales = "free") +
  theme_bw() +
  theme(legend.position = "none")
```

## Heatmap by continent{.tabset}
```{r}
# get contintents
dist_by_continent <- all_dyadic2 %>%
  rowwise() %>%
  mutate(country_name_1 = as.factor(countrycode(country_code_1, "iso2c","country.name")),
         country_name_2 = as.factor(countrycode(country_code_2, "iso2c","country.name")),
         continent_name_1 = countrycode(country_code_1, 'iso2c', 'continent'),
         continent_name_2 = countrycode(country_code_2, 'iso2c', 'continent')) %>%
  ungroup() %>%
  mutate(cont_order_1 = as.factor(continent_name_1) %>% as.numeric,
         cont_order_2 = as.factor(continent_name_2) %>% as.numeric) %>%
  select(-continent_name_1, -continent_name_2) %>%
  mutate(country_name_1 = fct_reorder(country_name_1, cont_order_1), 
         country_name_2 = fct_reorder(country_name_2, cont_order_2)) 
```

### centroid_dist_meters
```{r}
dist_by_continent %>%
  ggplot(aes(x = country_name_1, 
             y = country_name_2)) +
  geom_raster(aes(fill = centroid_dist_meters)) + 
  scale_fill_continuous(low = "#ffffcc", high = "#800026") +
  ggtitle("Geographical distance") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, 
                                   hjust = 1, vjust = 1, size = 6),
        axis.text.y = element_text(size = 6),
        axis.title = element_blank())
```

### wals_euclidean_dist
```{r,  fig.height = 8, fig.width = 10}
dist_by_continent %>%
  ggplot(aes(x = country_name_1, 
             y = country_name_2)) +
  geom_raster(aes(fill = wals_euclidean_dist)) + 
  scale_fill_continuous(low = "#ffffcc", high = "#800026") +
  ggtitle("WALS distance") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, 
                                   hjust = 1, vjust = 1, size = 6),
        axis.text.y = element_text(size = 6),
        axis.title = element_blank())
```

### asjp_dist
```{r,  fig.height = 8, fig.width = 10}
dist_by_continent %>%
  ggplot(aes(x = country_name_1, 
             y = country_name_2)) +
  geom_raster(aes(fill = asjp_dist)) + 
  scale_fill_continuous(low = "#ffffcc", high = "#800026") +
  ggtitle("ASJP distance") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, 
                                   hjust = 1, vjust = 1, size = 6),
        axis.text.y = element_text(size = 6),
        axis.title = element_blank())
```

### log_normalized_n_events_all
```{r,  fig.height = 8, fig.width = 10}
dist_by_continent %>%
  ggplot(aes(x = country_name_1, 
             y = country_name_2)) +
  geom_raster(aes(fill = log_normalized_n_events_all)) + 
  scale_fill_continuous(low = "#ffffcc", high = "#800026") +
  ggtitle("events") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, 
                                   hjust = 1, vjust = 1, size = 6),
        axis.text.y = element_text(size = 6),
        axis.title = element_blank())
```

### log_normalized_mean_imports_dollars
```{r,  fig.height = 8, fig.width = 10}
dist_by_continent %>%
  ggplot(aes(x = country_name_1, 
             y = country_name_2)) +
  geom_raster(aes(fill = log_normalized_mean_imports_dollars)) + 
  scale_fill_continuous(low = "#ffffcc", high = "#800026") +
  ggtitle("mean_imports_dollars") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, 
                                   hjust = 1, vjust = 1, size = 6),
        axis.text.y = element_text(size = 6),
        axis.title = element_blank())
```


## Between measure correlations
```{r, fig.height = 5, fig.width = 7}
corrs <- correlate(all_dyadic2 %>% select(-1:-3), 
                   use = "complete.obs")  %>%
  shave()

rplot(corrs, 
      legend = TRUE,
      colours = c("skyblue1", "white","indianred2")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

gather(corrs, 
       "variable1", "value", -1) %>%
  ungroup() %>%
  arrange(-value) %>%
  as.data.frame() %>%
  filter(!is.na(value)) %>%
  kable()
```

## Scatterplots{.tabset}

### dist vs. imports
```{r}
ggplot(all_dyadic2, aes(x = centroid_dist_meters,
                              y = log_normalized_mean_imports_dollars)) +
  geom_point() +
  geom_smooth(method = "lm")+
  theme_minimal()
```

### dist vs. events
```{r}
ggplot(all_dyadic2, aes(x = centroid_dist_meters,
                              y =  log_normalized_n_events_all)) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_minimal()
```

### events vs. imports
```{r}
ggplot(all_dyadic2, aes(x = log_normalized_n_events_all,
                              y = log_normalized_mean_imports_dollars)) +
  geom_point() +
  geom_smooth(method = "lm")+
  theme_minimal()
```

### dist vs. asjp
```{r}
ggplot(all_dyadic2, aes(x = centroid_dist_meters,
                              y = asjp_dist)) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_minimal()
```

### events vs. asjp
```{r}
ggplot(all_dyadic2, aes(x = log_normalized_n_events_all, 
                              y = asjp_dist)) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_minimal()
```

### imports vs. asjp
```{r}
ggplot(all_dyadic2, aes(x = log_normalized_mean_imports_dollars, 
                        y = asjp_dist)) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_minimal()
```

```{r}
#write_csv(dist_by_continent, "../../data/supplementary_data/cultural_sim_measures/all_dyadic_vars.csv")

```
