==> analyses/1_exploration/markdowns/10_compare_pairwise.Rmd <==
---
title: Compare two countries pairwise  similarities
author: Molly Lewis 
date: "`r Sys.Date()`"
output: 
  html_document:
    toc_float: false
    code_folding: hide
---
  
******
******
  
```{r setup, include = F}
rm(list=ls())

# load packages
library(knitr)
library(rmarkdown)
library(langcog)
library(tidyverse)
library(stringr)
library(purrr)
library(corrr)
library(countrycode)
library(feather)
library(forcats)
library(ggdendro)

opts_chunk$set(echo = T, message = F, warning = F, 
               error = F, tidy = F, cache = T, fig.height = 4)
```

```{r}
ITEM1 <- "tree"
ITEM2 <- "bread"
``` 

Tree vs. Bread
```{r}

get_unique_relation_id <- function (x, y){
  pairs = c(x, y)
  ordered = order(pairs)
  paste0(pairs[ordered[1]], pairs[ordered[2]])
}
d1 = read_csv(paste0("../../data/keras_similarities/pairwise_country/", ITEM1, "_sim_ratios.csv")) %>%
  mutate(item = ITEM1)
d2 = read_csv(paste0("../../data/keras_similarities/pairwise_country/", ITEM2, "_sim_ratios.csv"))%>%
  mutate(item = ITEM2)

alld <- bind_rows(d1, d2) %>%
  rowwise() %>%
  mutate(all_codes = get_unique_relation_id(country_code_1, country_code_2))  %>%
  ungroup() %>%
  select(item, mean_ratio, everything()) %>%
  spread(item, mean_ratio) %>%
  mutate(cont_order_1 =  as.factor(cont_order_1),
         cont_order_2 = as.factor(cont_order_2)) %>%
  mutate(cont_order_1 = fct_recode(cont_order_1, 
                                   "AM" = "2",
                                   "AU" = "5",
                                   "EU" = "4",
                                   "AS" = "3",
                                   "AF" = "1"),
         cont_order_2 = fct_recode(cont_order_2, 
                                   "AM" = "2",
                                   "AU" = "5",
                                   "EU" = "4",
                                   "AS" = "3",
                                   "AF" = "1")) 

  alld2 <- alld %>%
  group_by(all_codes) %>%
  summarize(tree = mean(tree, na.rm = T),
            bread = mean(bread, na.rm = T))

ggplot(alld2, aes(x = bread, y = tree )) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme_minimal()
```

```{r}

cont_means <- alld %>%
  group_by(cont_order_1, cont_order_2) %>%
  summarize(tree = mean(tree, na.rm = TRUE),
            bread = mean(bread, na.rm = TRUE))

ggplot(cont_means, aes(x = bread, y = tree )) +
  geom_point(aes(color = as.factor(cont_order_1), shape = as.factor(cont_order_2))) +
  geom_smooth(method = "lm") +
  theme_minimal()
```
