==> analyses/1_exploration/markdowns/14_d_place_exploration.Rmd <==
---
title: D_place_exploration
author: Molly Lewis 
date: "`r Sys.Date()`"
output: 
  html_document:
    toc_float: true
    code_folding: hide
---
  
******
******
  
```{r setup, include = F}
rm(list=ls())

# load packages
library(knitr)
library(rmarkdown)
library(tidyverse)
library(purrr)
library(feather)
library(forcats)
library(data.table)
library(countrycode)
library(dplacer)

source("../R_scripts/helpers.R")

opts_chunk$set(echo = T, message = F, warning = F, 
               error = F, tidy = F, cache = T, fig.height = 4)
```

Read in HD data
```{r}

countries <- read_csv("../../data/misc/major_countries.csv") %>%
  mutate(iso3 = tolower(iso_3),
          gc = gltc.iso(iso3),
          lang = lang.iso(iso3))

societies <- dplace_all_societies() 

societies_filt <- societies %>%
  select(c(1:2, 5, 7, 9,  11, 12, 13))

magrep <- function(x, y) {
  poss_agrep <- possibly(agrep, otherwise = 0)
  match_value <- poss_agrep(x,y)
  match_value <- ifelse(length(match_value) == 0, 0, match_value)
  ifelse(match_value > 0, T, F)
}

get_dplace_id <- function(country_name, iso3, gc, lang, societies){
 # print(country_name)
     m = societies %>%
        rowwise() %>%
        filter(#name == country_name|
              #alternate_names == country_name|
              #language.iso_code  == iso3 |
              #language.glotto_code == gc |
              #language.name == lang|
             # name == lang|
              magrep(name, country_name) |
              magrep(alternate_names, country_name) |
              magrep(language.name, lang) |
              magrep(name, lang)) %>%
    mutate(country_name = country_name) %>%
    select(country_name, everything())
}

p = pmap_df(select(countries, country_name, iso3, gc, lang),
        get_dplace_id, societies_filt) 
  


```
